Resolve review comments.
